Rename axis enumeration to api_axis_name
The enumeration "axis" is renamed such that the name "axis" to can be
used for a class representing axes. This class will follow in a future
commit. The new name also reflects that the enumeration represents the
name an axis may have in the API and thus as part of a request. The name
may not be present in the actual VDS dataset.