Add class to represent axes
A new class Axis is introduced to represent the axis of a dataset. It
contains basic information like number of points along an axis and the
min/max values and unit. Additionally, the axis knows its name,
coordinate system and (dimension) index according to the API definition
and the VDS definition.