Update Docker container base
Golang 1.22 is the currently most recent version of the Go runtime
environment. The Golang 1.20 runtime environment ran out of support on
6 Feb 2024.

Alpine 3.18 is the most recent version we can update to. Alpine 3.19
uses GCC 13 which is incompatible with the `azure-sdk-for-cpp` used by
OpenVDS 3.3.3. The issue is fixed in OpenVDS's master branch, but there
is no new release available yet.